use Arc instead of Arc.Definition
