+++ lib/arc.ex
@@ -1,2 +1,7 @@
 defmodule Arc do
+  defmacro __using__(_opts) do
+    quote do
+      use Arc.Definition
+    end
+  end
 end
